Add first version of findRepo script
First version of findRepo. It currently goes to git01 and tries to
find an exact match for the first argument given to the script.
If a repo is found, the clone url is written to stdout and the
script exits with status code 0. Otherwise it exits with status
code 1.

The SSL_CERT_FILE env variable had to be set so that ruby accepts
git01's certificate.